refactor: body and properties are now independent panels
