+++ menu/__init__.py
@@ -9,21 +9,23 @@ from . import (  # noqa: E402, I001
     menu_development,
     menu_warnings,
     menu_configure,
+    menu_configure_properties,
+    menu_configure_morphs,
     menu_configure_body,
     menu_configure_outfit,
     menu_configure_hair,
     menu_configure_armature,
     menu_configure_physics,
     menu_configure_tools,
-    menu_configure_morphs,
     menu_configure_links,
     menu_configure_others,
     menu_configure_debug,
     menu_configure_complete,
     menu_tools_creators,
     menu_quick_setup,
-    menu_body,
+    menu_properties,
     menu_morphs,
+    menu_body,
     menu_outfits,
     menu_hair,
     menu_armature,
@@ -41,21 +43,23 @@ def register():
     menu_development.register()
     menu_warnings.register()
     menu_configure.register()
+    menu_configure_properties.register()
+    menu_configure_morphs.register()
     menu_configure_body.register()
     menu_configure_outfit.register()
     menu_configure_hair.register()
     menu_configure_armature.register()
     menu_configure_physics.register()
     menu_configure_tools.register()
-    menu_configure_morphs.register()
     menu_configure_links.register()
     menu_configure_others.register()
     menu_configure_debug.register()
     menu_configure_complete.register()
     menu_tools_creators.register()
     menu_quick_setup.register()
-    menu_body.register()
+    menu_properties.register()
     menu_morphs.register()
+    menu_body.register()
     menu_outfits.register()
     menu_hair.register()
     menu_armature.register()
@@ -77,21 +81,23 @@ def unregister():
     menu_armature.unregister()
     menu_hair.unregister()
     menu_outfits.unregister()
-    menu_morphs.unregister()
     menu_body.unregister()
+    menu_morphs.unregister()
+    menu_properties.unregister()
     menu_quick_setup.unregister()
     menu_tools_creators.unregister()
     menu_configure_complete.unregister()
     menu_configure_debug.unregister()
     menu_configure_others.unregister()
     menu_configure_links.unregister()
-    menu_configure_morphs.unregister()
     menu_configure_tools.unregister()
     menu_configure_physics.unregister()
     menu_configure_armature.unregister()
     menu_configure_hair.unregister()
     menu_configure_outfit.unregister()
     menu_configure_body.unregister()
+    menu_configure_morphs.unregister()
+    menu_configure_properties.unregister()
     menu_configure.unregister()
     menu_warnings.unregister()
     menu_development.unregister()
