+++ menu/menu_body.py
@@ -4,114 +4,9 @@ from ..misc.geometry_nodes import (
     draw_geometry_nodes_modifier_inputs,
     geometry_nodes_modifier_inputs,
 )
-from ..misc.prop_utils import evaluate_rna
-from ..misc.ui_multiline import label_multiline
 from ..model_selection.active_object import mustardui_active_object
 from ..warnings.can_draw_ui import can_draw_ui
 from . import MainPanel
-
-
-def draw_section(
-    context,
-    layout,
-    obj,
-    settings,
-    rig_settings,
-    custom_props,
-    section,
-    section_id,
-    draw_sub=True,
-):
-    if rig_settings.body_custom_properties_name_order:
-        custom_properties_section = sorted(
-            [
-                x
-                for x in custom_props
-                if x.section == section.name
-                and not x.hidden
-                and (not x.advanced if not settings.advanced else True)
-            ],
-            key=lambda x: x.name,
-        )
-    else:
-        custom_properties_section = [
-            x
-            for x in custom_props
-            if x.section == section.name
-            and not x.hidden
-            and (not x.advanced if not settings.advanced else True)
-        ]
-
-    if (
-        len(custom_properties_section) > 0
-        and (not section.advanced or (section.advanced and settings.advanced))
-        and draw_sub
-    ):
-        box = layout
-
-        # Header
-        row = layout.row(align=False)
-        if section.collapsable:
-            row.prop(
-                section,
-                "collapsed",
-                icon="DOWNARROW_HLT" if not section.collapsed else "RIGHTARROW",
-                icon_only=True,
-                emboss=False,
-            )
-        if section.icon != "" and section.icon != "NONE":
-            row.label(text=section.name, icon=section.icon)
-        else:
-            row.label(text=section.name)
-        row.operator(
-            "mustardui.section_property_default", text="", icon="LOOP_BACK"
-        ).section_id = section_id
-
-        # Properties
-        if not section.collapsed:
-            box = layout.box()
-            if section.description != "":
-                box2 = box.box()
-                label_multiline(
-                    context=context,
-                    text=section.description,
-                    parent=box2,
-                    icon=section.description_icon,
-                )
-            for prop in custom_properties_section:
-                row = box.row()
-                if rig_settings.body_custom_properties_icons:
-                    row.label(text=prop.name, icon=prop.icon if prop.icon != "NONE" else "DOT")
-                else:
-                    row.label(text=prop.name)
-                if not prop.is_animatable:
-                    try:
-                        row.prop(evaluate_rna(prop.rna), prop.path, text="")
-                    except Exception:
-                        row.prop(
-                            settings,
-                            "custom_properties_error_nonanimatable",
-                            icon="ERROR",
-                            text="",
-                            icon_only=True,
-                            emboss=False,
-                        )
-                else:
-                    if prop.prop_name in obj.keys():
-                        row.prop(obj, f'["{prop.prop_name}"]', text="")
-                    else:
-                        row.prop(
-                            settings,
-                            "custom_properties_error",
-                            icon="ERROR",
-                            text="",
-                            icon_only=True,
-                            emboss=False,
-                        )
-
-        return box, not section.collapsed
-
-    return layout, False
 
 
 class PANEL_PT_MustardUI_Body(MainPanel, bpy.types.Panel):
@@ -127,34 +22,30 @@ class PANEL_PT_MustardUI_Body(MainPanel, bpy.types.Panel):
 
         res, arm = mustardui_active_object(context, config=0)
 
-        if arm is not None:
-            rig_settings = arm.MustardUI_RigSettings
-            custom_props = arm.MustardUI_CustomProperties
+        if arm is None:
+            return False
 
-            # Check if there is any property to show
-            prop_to_show = (
-                rig_settings.body_enable_subdiv
-                or rig_settings.body_enable_smoothcorr
-                or rig_settings.body_enable_solidify
-                or rig_settings.body_enable_material_normal_nodes
-                or rig_settings.body_enable_preserve_volume
-                or rig_settings.body_enable_geometry_nodes
+        rig_settings = arm.MustardUI_RigSettings
+
+        # Check if there is any property to show
+        prop_to_show = (
+            rig_settings.body_enable_subdiv
+            or rig_settings.body_enable_smoothcorr
+            or rig_settings.body_enable_solidify
+            or rig_settings.body_enable_material_normal_nodes
+            or rig_settings.body_enable_preserve_volume
+            or rig_settings.body_enable_geometry_nodes
+        )
+
+        # Check if geometry nodes support is active and there are geometry nodes
+        # on the body object
+        geometry_nodes_support = False
+        if rig_settings.model_body is not None and rig_settings.body_enable_geometry_nodes_support:
+            geometry_nodes_support = (
+                len([x for x in rig_settings.model_body.modifiers if x.type == "NODES"]) > 0
             )
 
-            # Check if geometry nodes support is active and there are geometry nodes
-            # on the body object
-            geometry_nodes_support = False
-            if (
-                rig_settings.model_body is not None
-                and rig_settings.body_enable_geometry_nodes_support
-            ):
-                geometry_nodes_support = (
-                    len([x for x in rig_settings.model_body.modifiers if x.type == "NODES"]) > 0
-                )
-
-            return res and (prop_to_show or len(custom_props) > 0 or geometry_nodes_support)
-
-        return False
+        return res and (prop_to_show or geometry_nodes_support)
 
     def draw(self, context):
 
@@ -162,7 +53,6 @@ class PANEL_PT_MustardUI_Body(MainPanel, bpy.types.Panel):
 
         poll, obj = mustardui_active_object(context, config=0)
         rig_settings = obj.MustardUI_RigSettings
-        custom_props = obj.MustardUI_CustomProperties
 
         layout = self.layout
 
@@ -223,94 +113,10 @@ class PANEL_PT_MustardUI_Body(MainPanel, bpy.types.Panel):
             row.scale_x = 0.7
             row.prop(rig_settings, "body_subdiv_rend_lv")
 
-        if len(custom_props) > 0:
-            unsorted_props = [x for x in custom_props if x.section == ""]
-            if len(unsorted_props) > 0:
-                box = layout.box()
-                box.label(text="Un-sorted properties", icon="LIBRARY_DATA_BROKEN")
-                for prop in [
-                    x
-                    for x in custom_props
-                    if x.section == ""
-                    and not x.hidden
-                    and (not x.advanced if not settings.advanced else True)
-                ]:
-                    row = box.row()
-                    if rig_settings.body_custom_properties_icons:
-                        row.label(
-                            text=prop.name,
-                            icon=prop.icon if prop.icon != "NONE" else "DOT",
-                        )
-                    else:
-                        row.label(text=prop.name)
-                    if not prop.is_animatable:
-                        try:
-                            row.prop(evaluate_rna(prop.rna), prop.path, text="")
-                        except Exception:
-                            row.prop(
-                                settings,
-                                "custom_properties_error_nonanimatable",
-                                icon="ERROR",
-                                text="",
-                                icon_only=True,
-                                emboss=False,
-                            )
-                    else:
-                        if prop.prop_name in obj.keys():
-                            row.prop(obj, f'["{prop.prop_name}"]', text="")
-                        else:
-                            row.prop(
-                                settings,
-                                "custom_properties_error",
-                                icon="ERROR",
-                                text="",
-                                icon_only=True,
-                                emboss=False,
-                            )
-
-            sec_num = len(rig_settings.body_custom_properties_sections)
-            id = 0
-            for section_id, section in enumerate(rig_settings.body_custom_properties_sections):
-                # Subsections are drawn inside standard sections
-                if section.is_subsection:
-                    continue
-
-                # Draw main section
-                sublayout, subcollapse = draw_section(
-                    context,
-                    layout,
-                    obj,
-                    settings,
-                    rig_settings,
-                    custom_props,
-                    section,
-                    section_id,
-                )
-
-                # Draw subsections if available
-                id = id + 1
-                if id >= sec_num:
-                    break
-
-                subsec = rig_settings.body_custom_properties_sections[id]
-                while subsec.is_subsection:
-                    draw_section(
-                        context,
-                        sublayout,
-                        obj,
-                        settings,
-                        rig_settings,
-                        custom_props,
-                        subsec,
-                        section_id,
-                        subcollapse,
-                    )
-                    id = id + 1
-                    if id >= sec_num:
-                        break
-                    subsec = rig_settings.body_custom_properties_sections[id]
-
         # Geometry nodes as sections
+        if rig_settings.model_body is None:
+            return
+
         gnm = [x for x in rig_settings.model_body.modifiers if x.type == "NODES"]
 
         if len(gnm) > 0 and rig_settings.body_enable_geometry_nodes_support:
