+++ menu/menu_configure_body.py
@@ -4,7 +4,6 @@ from .. import __package__ as base_package
 from ..model_selection.active_object import mustardui_active_object
 from ..warnings.can_draw_ui import can_draw_ui
 from . import MainPanel
-from .menu_configure import row_scale
 
 
 class PANEL_PT_MustardUI_InitPanel_Body(MainPanel, bpy.types.Panel):
@@ -28,7 +27,6 @@ class PANEL_PT_MustardUI_InitPanel_Body(MainPanel, bpy.types.Panel):
     def draw(self, context):
 
         layout = self.layout
-        scene = context.scene
 
         res, arm = mustardui_active_object(context, config=1)
         rig_settings = arm.MustardUI_RigSettings
@@ -52,113 +50,6 @@ class PANEL_PT_MustardUI_InitPanel_Body(MainPanel, bpy.types.Panel):
         col.separator()
         col.prop(rig_settings, "body_enable_preserve_volume")
         col.prop(rig_settings, "body_enable_material_normal_nodes")
-
-        # Sections
-        box = layout.box()
-        box.label(text="Sections", icon="LINENUMBERS_OFF")
-        if len(arm.MustardUI_CustomProperties) > 0:
-            row = box.row()
-            row.template_list(
-                "MUSTARDUI_UL_Section_UIList",
-                "The_List",
-                rig_settings,
-                "body_custom_properties_sections",
-                scene,
-                "mustardui_section_uilist_index",
-            )
-            col = row.column()
-            col2 = col.column(align=True)
-            col2.operator("mustardui.section_property_assign", text="", icon="PRESET")
-            col.separator()
-            col2 = col.column(align=True)
-            col2.operator("mustardui.section_add", text="", icon="ADD")
-            col2.operator("mustardui.section_delete", text="", icon="REMOVE")
-            col.separator()
-            col2 = col.column(align=True)
-            opup = col2.operator("mustardui.section_switch", icon="TRIA_UP", text="")
-            opup.direction = "UP"
-            opdown = col2.operator("mustardui.section_switch", icon="TRIA_DOWN", text="")
-            opdown.direction = "DOWN"
-
-            if (
-                scene.mustardui_section_uilist_index > -1
-                and len(rig_settings.body_custom_properties_sections) > 0
-            ):
-                sec = rig_settings.body_custom_properties_sections[
-                    scene.mustardui_section_uilist_index
-                ]
-
-                row = box.row()
-                row.label(text="Icon")
-                row.scale_x = row_scale
-                row.prop(sec, "icon", text="")
-
-                col = box.column(align=True)
-
-                row = col.row()
-                row.label(text="Description")
-                row.scale_x = row_scale
-                row.prop(sec, "description", text="")
-
-                row = col.row()
-                row.enabled = sec.description != ""
-                row.label(text="Icon")
-                row.scale_x = row_scale
-                row.prop(sec, "description_icon", text="")
-
-                col = box.column(align=True)
-                row = col.row()
-                row.enabled = scene.mustardui_section_uilist_index != 0
-                row.prop(sec, "is_subsection")
-
-                col = box.column(align=True)
-
-                row = col.row()
-                row.prop(sec, "advanced")
-
-                row = col.row()
-                row.prop(sec, "collapsable")
-        else:
-            box = box.box()
-            box.label(text="No property added yet", icon="ERROR")
-
-        # Custom properties
-        box = layout.box()
-        row = box.row()
-        row.label(text="Custom properties", icon="PRESET_NEW")
-        row.operator("mustardui.property_smartcheck", text="", icon="SHADERFX")
-        row.operator("mustardui.property_fix_path", text="", icon="DECORATE_DRIVER")
-
-        if len(arm.MustardUI_CustomProperties) > 0:
-            row = box.row()
-            row.template_list(
-                "MUSTARDUI_UL_Property_UIList",
-                "The_List",
-                arm,
-                "MustardUI_CustomProperties",
-                scene,
-                "mustardui_property_uilist_index",
-            )
-            col = row.column()
-            col.operator("mustardui.property_settings", icon="PREFERENCES", text="").type = "BODY"
-            col.separator()
-            col2 = col.column(align=True)
-            opup = col2.operator("mustardui.property_switch", icon="TRIA_UP", text="")
-            opup.direction = "UP"
-            opup.type = "BODY"
-            opdown = col2.operator("mustardui.property_switch", icon="TRIA_DOWN", text="")
-            opdown.direction = "DOWN"
-            opdown.type = "BODY"
-            col.separator()
-            col.operator("mustardui.property_remove", icon="X", text="").type = "BODY"
-
-            col = box.column(align=True)
-            col.prop(rig_settings, "body_custom_properties_icons")
-            col.prop(rig_settings, "body_custom_properties_name_order")
-
-        else:
-            box = box.box()
-            box.label(text="No property added yet", icon="ERROR")
 
 
 def register():
